+++ wahlRaumFinder/store/index.js
@@ -1,10 +1,12 @@
 import getters from "./gettersWahlRaumFinder";
 import mutations from "./mutationsWahlRaumFinder";
+import actions from "./actionsWahlRaumFinder";
 import state from "./stateWahlRaumFinder";
 
 export default {
     namespaced: true,
     state: {...state},
     mutations,
+    actions,
     getters
 };